Stabilized SDE and BSP installation.
